rename api to functions
